Phase C-1 cascade: move ResumeFrameLayoutSummary to majit-ir::resumedata
Per-frame layout summary used by ResumeLayoutSummary.  All field types
(Type, ResumeValueKind, ResumeValueLayoutSummary) are already in
majit-ir, so the struct moves cleanly.

Three inherent impls (to_frame_info, to_exit_frame_layout, the public
from_exit_frame_layout constructor) reference FrameInfo / ExitFrameLayout
in majit-backend and the ResumeValueLayoutSummaryExt trait — extracted
as free functions in metainterp::resume:
- resume_frame_layout_to_frame_info
- resume_frame_layout_to_exit_frame_layout
- resume_frame_layout_from_exit_frame_layout (pub)

Two compile.rs callsites updated to free-fn syntax.
